// Neptune.Web/src/app/shared/generated/model/treatment-bmp-location-update.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPLocationUpdate { 
    Latitude: number;
    Longitude: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPLocationUpdateForm { 
    Latitude: FormControl<number>;
    Longitude: FormControl<number>;
}

export class TreatmentBMPLocationUpdateFormControls { 
    public static Latitude = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static Longitude = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
}